requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-restframework==0.0.1
drf-yasg==1.21.7
pillow==10.4.0
psycopg2==2.9.9